Fix not being able to delete bypass_actors
This is a quirk upstream, where they introduced a new function
`UpdateRulesetNoBypassActor` to work around the issue of
`UpdateRuleset` not deleting BypassActors when there's an empty array.
https://github.com/google/go-github/pull/3174

Not sure why they decided to introduce a new function instead of just
patching the original. Compatibility issues?
Is there ever a use case where we need to keep BypassActors around
when the non-empty list becomes an empty list?